added HaploidWriter walker
Added HaploidWriter walker to the phasing module. It is needed by PhaseByTransmission for processing X chromosome data.

It transforms diploid genotype calls into haploid genotype calls based on the previously estimated phred-scaled (diploid) likelihoods, without re-estimating them from the alignment data (either not-available or simply for conveniance). It is a standalone walker, but all the functonality is implemented as static so that it can easily be called by other tools/walkers (i.e.: PhaseByTransmission).

We can additionaly show a figure in which we compare the phred-scaled haploid genotype likelihoods estimated from the data by UnifiedGenotyper and the HaploidWriter re-computed haploid genotype likelihoods, and diferences observed are not meaningful.